long long -> long

+++ problems/029_campionatul/mihania.cpp
@@ -3,11 +3,11 @@
 #include<numeric>
 using namespace std;
 
-vector<long long> getSums(long long S, vector<long long>& c) {
-	vector<long long> res;
-	long long max = (long long)pow(2, c.size());
-	for (long long k = 0; k < max; k++) {
-		long long sum = 0;
+vector<long> getSums(long S, vector<long>& c) {
+	vector<long> res;
+	long max = (long)pow(2, c.size());
+	for (long k = 0; k < max; k++) {
+		long sum = 0;
 		int idx = 0;
 		auto mask = k;
 		while (mask > 0) {
@@ -28,11 +28,11 @@ vector<long long> getSums(long long S, vector<long long>& c) {
 	return res;
 }
 
-long long solve(long long S, vector<long long>& c) {
+long solve(long S, vector<long>& c) {
 	
 	// splitting array into two equal subarrays
-	auto v1 = vector<long long>(c.begin(), c.begin() + c.size() / 2);
-	auto v2 = vector<long long>(c.begin() + c.size() / 2, c.end());
+	auto v1 = vector<long>(c.begin(), c.begin() + c.size() / 2);
+	auto v2 = vector<long>(c.begin() + c.size() / 2, c.end());
 
 	// calculating all possible subset sums in each array
 	auto s1 = getSums(S, v1);
@@ -44,7 +44,7 @@ long long solve(long long S, vector<long long>& c) {
 	s2.erase(s2.begin());
 
 	// merging arrays subset sums
-	long long res = 0;
+	long res = 0;
 	for (auto k : s1) {
 		auto it = upper_bound(s2.begin(), s2.end(), S - k);
 		res += 1 + (it - s2.begin());
@@ -60,13 +60,13 @@ int main() {
 
 	ofstream out("tests.out");
 	for (auto i = 0; i < T; i++) {
-		long long N;
-		long long S;
-		vector<long long> c;
+		long N;
+		long S;
+		vector<long> c;
 		in >> N;
 		in >> S;
 		for (auto j = 0; j < N; j++) {
-			long long v;
+			long v;
 			in >> v;
 			c.push_back(v);
 		}
